feat: return undefined if no resolver is matched
Previously nestedMap would return the value itself if no resolver was found for the value. This
behaviour assumed that the library's users intent would be to keep the value if no resolver was
found, instead of removing it. This commit modifies that behaviour in order to allow library's users
to explicitly define the default behaviour through a default resolver alternative.

BREAKING CHANGE: return undefined if no resolver is matched instead of original value